+++ dune/localfunctions/lagrange/qksubsampled/qksubsampledlocalinterpolation.hh
@@ -6,11 +6,15 @@
 
 #include <dune/common/fvector.hh>
 #include <dune/common/power.hh>
+#include <dune/common/version.hh>
 
 #include <dune/geometry/type.hh>
 
 #include <dune/localfunctions/common/localbasis.hh>
 #include <dune/localfunctions/common/localfiniteelementtraits.hh>
+#if DUNE_VERSION_NEWER(DUNE_LOCALFUNCTIONS,2,7)
+#include <dune/localfunctions/common/localinterpolation.hh>
+#endif
 
 
 namespace Dune
@@ -36,10 +40,15 @@ namespace Dune
 
     //! \brief Local interpolation of a function
     template<typename F, typename C>
-    void interpolate (const F& f, std::vector<C>& out) const
+    void interpolate (const F& ff, std::vector<C>& out) const
     {
       typename LB::Traits::DomainType x;
+
+#if DUNE_VERSION_NEWER(DUNE_LOCALFUNCTIONS,2,7)
+      auto&& f = Impl::makeFunctionWithCallOperator<typename LB::Traits::DomainType>(ff);
+#else
       typename LB::Traits::RangeType y;
+#endif
 
       out.resize(StaticPower<s*k+1,d>::power);
 
@@ -52,7 +61,12 @@ namespace Dune
         for (int j=0; j<d; j++)
           x[j] = (1.0*alpha[j])/(s*k);
 
-        f.evaluate(x,y); out[i] = y;
+#if DUNE_VERSION_NEWER(DUNE_LOCALFUNCTIONS,2,7)
+        out[i] = f(x);
+#else
+        ff.evaluate(x,y);
+        out[i] = y;
+#endif
       }
     }
   };
@@ -64,13 +78,23 @@ namespace Dune
   public:
     //! \brief Local interpolation of a function
     template<typename F, typename C>
-    void interpolate (const F& f, std::vector<C>& out) const
+    void interpolate (const F& ff, std::vector<C>& out) const
     {
       typename LB::Traits::DomainType x(0);
+
+#if DUNE_VERSION_NEWER(DUNE_LOCALFUNCTIONS,2,7)
+      auto&& f = Impl::makeFunctionWithCallOperator<typename LB::Traits::DomainType>(ff);
+#else
       typename LB::Traits::RangeType y;
-      f.evaluate(x,y);
+#endif
+
       out.resize(1);
+#if DUNE_VERSION_NEWER(DUNE_LOCALFUNCTIONS,2,7)
+      out[0] = f(x);
+#else
+      ff.evaluate(x,y);
       out[0] = y;
+#endif
     }
   };